Implement basic Mock API in new build flavor
