+++ Corona-Warn-App/build.gradle
@@ -87,8 +87,15 @@ android {
     }
 
     // One version contains our Test Fragments
-    flavorDimensions "version"
+    flavorDimensions "version", "api"
     productFlavors {
+        sdkApi {
+            dimension "api"
+        }
+        mockApi {
+            dimension "api"
+            applicationIdSuffix '.mockapi'
+        }
         device {
             dimension "version"
             buildConfigField "String", "BUILD_VARIANT", "\"device\""
@@ -165,6 +172,16 @@ android {
     }
 
     sourceSets {
+        sdkApi {
+            kotlin {
+                srcDirs = ['src/sdkApi']
+            }
+        }
+        mockApi {
+            kotlin {
+                srcDirs = ['src/mockApi']
+            }
+        }
         deviceForTesters {
             kotlin {
                 srcDirs = ['src/deviceForTesters']
@@ -219,6 +236,10 @@ task jacocoTestReportDeviceRelease(type: JacocoReport, dependsOn: 'testDeviceRel
     getExecutionData().from(fileTree(dir: "$buildDir", includes: ["jacoco/testDeviceReleaseUnitTest.exec"]))
 }
 
+configurations.all {
+    resolutionStrategy.cacheChangingModulesFor 30, 'seconds'
+}
+
 dependencies {
     // KOTLIN
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:1.3.7"
@@ -263,8 +284,9 @@ dependencies {
     debugImplementation 'androidx.fragment:fragment-testing:1.2.4'
 
     // CoraLibre sdk including the FOSS EN service reimplementation
-    implementation('com.github.CoraLibre:CoraLibre-android-sdk:592ec0093d08a0fb276150c45ecfe309f4a694d9') {
+    implementation('com.github.CoraLibre:CoraLibre-android-sdk:master-SNAPSHOT') {
         exclude group: 'com.google.protobuf', module: 'protobuf-java'
+        changing = true
     }
 
     // HTTP
