+++ Corona-Warn-App/src/sdkApi/java/de.rki.coronawarnapp/nearby/ExposureNotificationClientFactory.kt
@@ -2,9 +2,10 @@ package de.rki.coronawarnapp.nearby
 
 import de.rki.coronawarnapp.CoronaWarnApplication
 import org.coralibre.android.sdk.fakegms.nearby.Nearby
+import org.coralibre.android.sdk.fakegms.nearby.exposurenotification.ExposureNotificationClient
 
 object ExposureNotificationClientFactory {
-    fun createClient(): Nearby {
+    fun createClient(): ExposureNotificationClient {
         return Nearby.getExposureNotificationClient(CoronaWarnApplication.getAppContext())
     }
 }
